requirements.txt
opencv-python
pytesseract
pyautogui
Pillow
